Add files via upload
This is a substantial change set.  A the request of reviewers, I've changed the terminology of SECONDARY_SERVERS and TERTIARY_SERVERS to SYNC_SERVERS and ASYNC_SERVERS to better align with SQL Server terminology.  There is also a new ag-add.sh script that simplifies adding one or more nodes of a given type to an existing cluster defined in the params.sh file. There is a script to install RHEL HA related software on newly provisioned nodes as well.